move the election to pkg
