+++ server/tso/allocator_manager.go
@@ -28,13 +28,13 @@ import (
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/kvproto/pkg/pdpb"
 	"github.com/pingcap/log"
+	"github.com/tikv/pd/pkg/election"
 	"github.com/tikv/pd/pkg/errs"
 	"github.com/tikv/pd/pkg/etcdutil"
 	"github.com/tikv/pd/pkg/grpcutil"
 	"github.com/tikv/pd/pkg/slice"
 	"github.com/tikv/pd/pkg/syncutil"
 	"github.com/tikv/pd/server/config"
-	"github.com/tikv/pd/server/election"
 	"github.com/tikv/pd/server/member"
 	"github.com/tikv/pd/server/storage/kv"
 	"go.etcd.io/etcd/clientv3"
